documentation

+++ form4.php
@@ -7,21 +7,24 @@ error_reporting(E_ALL ^ E_NOTICE ^ E_WARNING);
 	
 	$usrid1 = $_SESSION['userid'];
 
+	// to check whether the form is submitted or not
 	if(mysqli_num_rows(mysqli_query($con, "select submitted from form4 where userid = $usrid1 and submitted = 1")) > 0)
 	{
-		echo "<br/><br/><br/><br/>Your form is already submitted<br>";
-		echo 'Click <a href="pdf_final.php">here</a></li> to generate the pdf of your application<br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>';
+		echo "<br/><br/><br/><br/>Your form is already submitted<br><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>";
+		//echo 'Click <a href="pdf_final.php">here</a></li> to generate the pdf of your application<br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>';
         include 'footer.php';
 
 		//echo '<p><a href=".">Home</a></p>';
 		exit;
 	}
+	//end of checking
+
 	if($_GET[a]==1)
 	{
 		echo "<br/><span style='color:green;'>Professional Activities SAVED</span>";
 	}
 
-
+	// retrieving data from form4 table 
 	$retrieve = "SELECT * FROM form4 where userid = $usrid1";
 	$result = mysqli_query($con,$retrieve);
 	while($row = mysqli_fetch_array($result))
@@ -44,12 +47,12 @@ error_reporting(E_ALL ^ E_NOTICE ^ E_WARNING);
 		$othr=$row['otherinfo27'];
 	
   	}
+  	//end of retrieval
 
 if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1])) 
 {
 
-	echo "<br/><span style='color:green;'>SOP and LOR SAVED</span><br/>";
-	
+		//copying submitted form variables		
 		$ref1_name=$_REQUEST['ref1_name'];
 		$ref1_add=$_REQUEST['ref1_add'];
 		$ref1_email=$_REQUEST['ref1_email'];
@@ -63,7 +66,31 @@ if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1]))
 		$ref3_email=$_REQUEST['ref3_email'];
 		$ref3_phone=$_REQUEST['ref3_phone'];
 		$othr=$_REQUEST['27'];
-	/*	$errorMessage = "";
+		//end of copying
+
+		//validation of referee emails with user email
+		$errorMessage = "";
+		$retrieve = "SELECT email FROM users where id_user = $usrid1";
+		$result = mysqli_query($con,$retrieve);
+		while($row = mysqli_fetch_array($result))
+	  	{
+			$usermail=$row['email'];
+	  	}
+	  	if($usermail==$ref1_email)
+	  	{
+	  		$errorMessage .= "<li>check the referee1 email. It is similar to your email</li>";
+	  	}
+	  	if($usermail==$ref2_email)
+	  	{
+			$errorMessage .= "<li>check the referee2 email. It is similar to your email</li>";	  		
+	  	}
+	  	if($usermail==$ref3_email)
+	  	{
+	  		$errorMessage .= "<li>check the referee3 email. It is similar to your email</li>";
+	  	}
+	  	//end of validation
+
+		/*	$errorMessage = "";
 		
 		if(empty($_POST['25a'] )) 
         	{
@@ -85,7 +112,7 @@ if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1]))
 	        {
 			$errorMessage .= "<li>You have forgotten to mention your Referee-3's Details !</li>";
 		}		
-
+		
 			
 		if(empty($errorMessage)) 
         	{ 
@@ -169,16 +196,25 @@ if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1]))
 		// End upload script
 	}
 
-		
+		if($errorMessage == "")
+		{
+
+		//writing the validated data in to the form4 table			
+		echo "<br/><span style='color:green;'>SOP and LOR SAVED</span><br/>";			
 		mysqli_query($con,"delete from form4 where userid=$usrid1");
 		$sql="INSERT INTO form4(userid,sop25a,sop25b,ref1_name,ref1_addr,ref1_email,ref1_phone,ref2_name,ref2_addr,ref2_email,ref2_phone,ref3_name,ref3_addr,ref3_email,ref3_phone,otherinfo27,submitted) VALUES ('$usrid1','$a25','$b25','$ref1_name','$ref1_add','$ref1_email','$ref1_phone','$ref2_name','$ref2_add','$ref2_email','$ref2_phone','$ref3_name','$ref3_add','$ref3_email','$ref3_phone','$othr',0)";
      	mysqli_query($con,$sql);
+     	//end of writing
 
 		if(isset($_POST[submitted_val1]))
 			{	
 			echo '<meta http-equiv="REFRESH" content="0;url=submit.php?a=1">';
 			}
-	
+		}
+		else
+		{
+			echo "<ul>".$errorMessage."</ul>";
+		}
 
    /* if(!empty($errorMessage)) 
 		    	{
@@ -192,6 +228,8 @@ if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1]))
 
    ?>
    <script type="text/javascript">
+   
+   // function to check the file size and extension
 	function check_25(a){
                 str=document.getElementById(a).value.toUpperCase();
         suffix=".PDF";
@@ -206,91 +244,100 @@ if(isset($_POST[submitted_val]) || isset($_POST[submitted_val1]))
             document.getElementById(a).value='';        		        		
         }        
     }
+    //end of file checking
+
     </script>
 
 <html>
-<body background="bgimage.jpg";background-repeat:no-repeat;background-attachment:scroll;>
+	<body background="bgimage.jpg";background-repeat:no-repeat;background-attachment:scroll;>
 		<div class="row">
-
 			<div class="col-md-12">
-
 				<form method="post" class="form" action="form4.php" enctype="multipart/form-data">
-<br/>
-<span style="color:red;">* required fields</span>
-<br/><br/>
-<table class="table table-striped" id="myTable">
+					<br/>
+					<span style="color:red;">* required fields</span>
+					<br/><br/>
 
-25:SOP<span style="color:red;">*</span><br/><br/>
-<tr>
-<td>a) Why would you like to join IIITDM Kancheepuram? (file size < 2MB is accepted.)<br/></td>
-<td>b) Your vision for the growth of the institute...  (file size < 2MB is accepted.)<br/></td>
-<!--<td><textarea class="form-control" style= "width: 500px; height: 150px;" row="20" column="200" name="25a" maxlength="4000"><?php echo $a25;?></textarea></td> -->
-</tr>
-<tr>
-<!-- <td><textarea class="form-control" style="width: 500px; height: 150px;" row="20" column="200" name="25b" maxlength="4000"><?php echo $b25;?></textarea></td> -->
-<td><input type="file" name="25a" id="25a" onchange="check_25('25a')"><?php if(strlen($a25) > 0) echo "It is already submitted; To overwrite, upload again"; ?><br/></td>
-<td><input type="file" name="25b" id="25b" onchange="check_25('25b')"><?php if(strlen($b25) > 0) echo "It is already submitted; To overwrite, upload again"; ?><br/></td>
-</tr>
-</table>
+					<!-- file fields for uploading SOPs -->
+					<table class="table table-striped" id="myTable">
 
-<table class="table table-striped" id="myTable">
+						25:SOP<span style="color:red;">*</span><br/><br/>
+						<tr>
+							<td>a) Why would you like to join IIITDM Kancheepuram? (file size < 2MB is accepted.)<br/></td>
+							<td>b) Your vision for the growth of the institute...  (file size < 2MB is accepted.)<br/></td>
+							<!--<td><textarea class="form-control" style= "width: 500px; height: 150px;" row="20" column="200" name="25a" maxlength="4000"><?php echo $a25;?></textarea></td> -->
+						</tr>
+						<tr>
+							<!-- <td><textarea class="form-control" style="width: 500px; height: 150px;" row="20" column="200" name="25b" maxlength="4000"><?php echo $b25;?></textarea></td> -->
+							<td><input type="file" name="25a" id="25a" onchange="check_25('25a')"><?php if(strlen($a25) > 0) echo "It is already submitted; To overwrite, upload again"; ?><br/></td>
+							<td><input type="file" name="25b" id="25b" onchange="check_25('25b')"><?php if(strlen($b25) > 0) echo "It is already submitted; To overwrite, upload again"; ?><br/></td>
+						</tr>
+					</table>
+					<!-- end of SOPs -->
 
-26:Referee Details<span style="color:red;">*</span><br/>
-Enter the names and addresses including email,fax, telephone no. of 3 referees.<br/>
-(at least one of them should be familiar with your recent work);<br/>Referees will be contacted by the institute directly, if required.<br/><br/>
-<tr>
-<td><h4>Referee1:<br/></h4></td>
-<tr><td>Name:</td><td><input class="form-control" type="text" name="ref1_name" value="<?php echo $ref1_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
-<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref1_add" value="<?php echo $ref1_add ?>" style="width:50%;"/></td></tr>
-<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref1_email" value="<?php echo $ref1_email?>" style="width:50%;"/></td></tr>
-<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref1_phone" value="<?php echo $ref1_phone ?>" min="0" style="width:50%;"/></td></tr>
-</tr>
-<tr>
-<td><h4>Referee2:</h4></td>
-<tr><td>Name:</td><td><input class="form-control" type="text" name="ref2_name" value="<?php echo $ref2_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
-<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref2_add" value="<?php echo $ref2_add ?>" style="width:50%;"/></td></tr>
-<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref2_email" value="<?php echo $ref2_email?>" style="width:50%;"/></td></tr>
-<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref2_phone" value="<?php echo $ref2_phone ?>" min="0" style="width:50%;"/></td></tr>
-</tr>
-<tr>
-<td><h4>Referee3:</h4>
-<tr><td>Name:</td><td><input class="form-control" type="text" name="ref3_name" value="<?php echo $ref3_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
-<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref3_add" value="<?php echo $ref3_add ?>" style="width:50%;"/></td></tr>
-<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref3_email" value="<?php echo $ref3_email?>" style="width:50%;"/></td></tr>
-<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref3_phone" value="<?php echo $ref3_phone ?>" min="0" style="width:50%;"/></td></tr>
-</tr>
-</table>
-<!--
-<br/>
-Referee2:<br/>
-Name: <input type="text" name="ref2_name" value="<?php echo $ref2_name  ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters"/><br/>
-Address: <input type="text" name="ref2_add" value="<?php echo $ref2_add ?>" /><br/>
-Email: <input type="email" name="ref2_email" value="<?php echo $ref2_email ?>" /><br/>
-Phone: <input type="number" name="ref2_phone" value="<?php echo $ref2_phone ?>" min="0"/><br/>
-<br/><br/>
-Referee3:<br/>
-Name: <input type="text" name="ref3_name" value="<?php echo $ref3_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters"/><br/>
-Address: <input type="text" name="ref3_add" value="<?php echo $ref3_add ?>" /><br/>
-Email: <input type="email" name="ref3_email" value="<?php echo  $ref3_email?>" /><br/>
-Phone: <input type="number" name="ref3_phone" value="<?php echo $ref3_phone ?>" min="0"/><br/>
-</table>
-<br/>
--->
-<table class="table table-striped" id="myTable">
-<tr>
-<td>
-27:
-Any other information you want to mention:</td>
-<td><textarea class="form-control" style="width: 900px; height: 150px; " row="20" column="200" name="27"><?php echo $othr;?></textarea></td></tr></table>
-						<div class="text-center">
-							<input type="submit" class="btn btn-sm btn-info" name = "submitted_val" value="Save">
-							<input type="submit" class="btn btn-sm btn-success" name = "submitted_val1" value="Save & Next">
-						</div>
-</form>
-</h3>
-</div>
-</div>
-</body>
+					<!-- Referee details -->
+					<table class="table table-striped" id="myTable">
+
+						26:Referee Details<span style="color:red;">*</span><br/>
+						Enter the names and addresses including email,fax, telephone no. of 3 referees.<br/>
+						(at least one of them should be familiar with your recent work);<br/>Referees will be contacted by the institute directly, if required.<br/><br/>
+						<tr>
+							<td><h4>Referee1:<br/></h4></td>
+							<tr><td>Name:</td><td><input class="form-control" type="text" name="ref1_name" value="<?php echo $ref1_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
+							<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref1_add" value="<?php echo $ref1_add ?>" style="width:50%;"/></td></tr>
+							<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref1_email" value="<?php echo $ref1_email?>" style="width:50%;"/></td></tr>
+							<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref1_phone" value="<?php echo $ref1_phone ?>" min="0" style="width:50%;"/></td></tr>
+						</tr>
+						<tr>
+							<td><h4>Referee2:</h4></td>
+							<tr><td>Name:</td><td><input class="form-control" type="text" name="ref2_name" value="<?php echo $ref2_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
+							<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref2_add" value="<?php echo $ref2_add ?>" style="width:50%;"/></td></tr>
+							<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref2_email" value="<?php echo $ref2_email?>" style="width:50%;"/></td></tr>
+							<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref2_phone" value="<?php echo $ref2_phone ?>" min="0" style="width:50%;"/></td></tr>
+						</tr>
+						<tr>
+							<td><h4>Referee3:</h4>
+							<tr><td>Name:</td><td><input class="form-control" type="text" name="ref3_name" value="<?php echo $ref3_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters" style="width:50%;"/></td></tr>
+							<tr><td>Address:</td> <td><input class="form-control" type="text" name="ref3_add" value="<?php echo $ref3_add ?>" style="width:50%;"/></td></tr>
+							<tr><td>Email:</td> <td><input class="form-control" type="email" name="ref3_email" value="<?php echo $ref3_email?>" style="width:50%;"/></td></tr>
+							<tr><td>Phone:</td> <td><input class="form-control" type="number" name="ref3_phone" value="<?php echo $ref3_phone ?>" min="0" style="width:50%;"/></td></tr>
+						</tr>
+					</table>
+					<!-- end of Referee details -->
+
+					<!--
+					<br/>
+					Referee2:<br/>
+					Name: <input type="text" name="ref2_name" value="<?php echo $ref2_name  ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters"/><br/>
+					Address: <input type="text" name="ref2_add" value="<?php echo $ref2_add ?>" /><br/>
+					Email: <input type="email" name="ref2_email" value="<?php echo $ref2_email ?>" /><br/>
+					Phone: <input type="number" name="ref2_phone" value="<?php echo $ref2_phone ?>" min="0"/><br/>
+					<br/><br/>
+					Referee3:<br/>
+					Name: <input type="text" name="ref3_name" value="<?php echo $ref3_name ?>" pattern='[a-zA-Z0-9]{0,100}' title="Only alphanumeric input is valid upto 100 characters"/><br/>
+					Address: <input type="text" name="ref3_add" value="<?php echo $ref3_add ?>" /><br/>
+					Email: <input type="email" name="ref3_email" value="<?php echo  $ref3_email?>" /><br/>
+					Phone: <input type="number" name="ref3_phone" value="<?php echo $ref3_phone ?>" min="0"/><br/>
+					</table>
+					<br/>
+					-->
+
+					<!-- textarea tag for entering any other information -->
+					<table class="table table-striped" id="myTable">
+						<tr>
+							<td>27:Any other information you want to mention:</td>
+							<td><textarea class="form-control" style="width: 900px; height: 150px; " row="20" column="200" name="27"><?php echo $othr;?></textarea></td>
+						</tr>
+					</table>
+					<!-- end of any other information field -->
+
+					<div class="text-center">
+						<input type="submit" class="btn btn-sm btn-info" name = "submitted_val" value="Save">
+						<input type="submit" class="btn btn-sm btn-success" name = "submitted_val1" value="Save & Next">
+					</div>
+				</form>
+			</div>
+		</div>
+	</body>
 </html>
 <?php include 'footer.php'; ?>
